// src/zap-data/types/enums.ts
// Enum for project types
export enum Project {
    ApeBond = 'ApeBond',
    Uniswap = 'Uniswap',
    PancakeSwap = 'PancakeSwap',
    SushiSwap = 'SushiSwap',
    QuickSwap = 'QuickSwap',
    Zyberswap = 'Zyberswap',
    Thena = 'Thena',
    Retro = 'Retro',
    Ascent = 'Ascent',
    Ramses = 'Ramses',
    Camelot = 'Camelot',
    Lynex = 'Lynex',
    SynthSwap = 'SynthSwap',
    BaseX = 'BaseX',
    SwapBased = 'SwapBased',
}